XORLearn

+++ CSharpBible/Graphics/XORLearn/Program.cs
@@ -30,7 +30,7 @@ class Program
         NeuralNetwork nn = IoC.GetRequiredService<NeuralNetwork>();
 
         // 2. Trainingsdaten (XOR-Logik)47
-        double[][] inputs =
+        float[][] inputs =
         [
             [0, 0,0],
             [0, 1 ,0],
@@ -42,7 +42,7 @@ class Program
             [1, 1,1]
         ];
 
-        double[][] targets =
+        float[][] targets =
         [
             [0,0],
             [1,0],
@@ -71,7 +71,7 @@ class Program
                 var result = 1.0;
                 foreach (var tuple in inputs.Zip(targets))
                 {
-                    double[] output = nn.FeedForward(tuple.First);
+                    float[] output = nn.FeedForward(tuple.First);
                     result *= (output[0] - 0.5) * (tuple.Second[0] - 0.5) * 4;
                     result *= (output[1] - 0.5) * (tuple.Second[1] - 0.5) * 4;
                 }
@@ -85,7 +85,7 @@ class Program
         var finalresult = 1.0;
         foreach (var tuple in inputs.Zip(targets))
         {
-            double[] output = nn.FeedForward(tuple.First);
+            float[] output = nn.FeedForward(tuple.First);
             Console.WriteLine($"Input: {tuple.First[0]}, {tuple.First[1]}, {tuple.First[2]} | Vorhersage: {output[0]:F3}, {output[1]:F3}");
             finalresult *= (output[0] - 0.5) * (tuple.Second[0] - 0.5) * 4;
             finalresult *= (output[1] - 0.5) * (tuple.Second[1] - 0.5) * 4;
